+++ ribs/emitters/opt/_cma_es.py
@@ -158,10 +158,9 @@ class CMAEvolutionStrategy:
         if area < 1e-11:
             return True
 
-        # Fitness is too flat (only applies if there are at least 2
-        # parents).
+        # Fitness is too flat (only applies if there are at least 2 parents).
         if (len(ranking_values) >= 2 and
-                np.abs(ranking_values[0] - ranking_values[-1]) < 1e-12):
+                np.linalg.norm(ranking_values[0] - ranking_values[-1]) < 1e-12):
             return True
 
         return False