+++ HISTORY.md
@@ -6,11 +6,12 @@
 
 #### API
 
+- Implement `EvolutionStrategyEmitter` (#220)
 - Raise ValueError for incorrect array shapes in archive methods (#219)
 - Add elites_with_measures_single method for getting elite for a single
   solution's measures (#215)
 - Introduced the Ranker object, which is responsible for ranking the solutions
-  based on different objectives (#209)
+  based on different objectives (#209, #222)
 - Add index_of_single method for getting index of measures for one solution
   (#214)
 - **Backwards-incompatible:** Replace elite_with_behavior with batched